requirements.txt
cherrypy==16.0.2
schedule==0.5.0
paho-mqtt==1.3.1
RPi.GPIO==0.6.3
